add authentication
Adds authentication for create/update/delete operations on posts.

This is a single-user app with no ambitions of becoming anything else,
so basic http auth over https covers the needs here.

Instead of the full set of CRUD actions and views for both public posts
and admin posts, the actions and views are split. Public posts -- which
are named PublishedPosts -- get `:show` and `:index`, while admin posts
(in `/admin/posts`) get the CUD. This keeps the flow simple and
unsurprising to the user.